--- SharePointFramework/PortfolioWebParts/src/components/ProjectProvision/ProjectProvision.module.scss
@@ -1,45 +1,0 @@
-.toolbar {
-  justify-content: space-between;
-}
-
-.body {
-  flex: 1;
-  width: 100%;
-  max-width: 100%;
-  position: relative;
-
-  .content {
-    display: flex;
-    flex-direction: column;
-    gap: 16px;
-
-    .sitetypes {
-      display: flex;
-      gap: 12px;
-      flex-wrap: wrap;
-    }
-
-    .ignoreGap {
-      margin-top: 0;
-      margin-bottom: 0;
-    }
-
-    .footer {
-      margin: 12px 0 0 0;
-    }
-
-    .flexContent {
-      display: flex;
-      flex-direction: row;
-      gap: 16px;
-    }
-  }
-}
-
-.footer {
-  justify-content: space-between;
-}
-
-div[role="dialog"] {
-  min-width: 1120px;
-}
