+++ SharePointFramework/PortfolioWebParts/src/components/ProjectProvision/ProvisionDrawer/ProvisionDrawer.module.scss
@@ -1,0 +1,57 @@
+.toolbar {
+  justify-content: space-between;
+}
+
+.body {
+  flex: 1;
+  width: 100%;
+  max-width: 100%;
+  position: relative;
+
+  .content {
+    display: flex;
+    flex-direction: column;
+    gap: 16px;
+
+    .sitetypes {
+      display: flex;
+      gap: 12px;
+      flex-wrap: wrap;
+    }
+
+    .ignoreGap {
+      margin-top: 0;
+      margin-bottom: 0;
+    }
+
+    .footer {
+      margin: 12px 0 0 0;
+    }
+
+    .flexContent {
+      display: flex;
+      flex-direction: row;
+      gap: 16px;
+    }
+  }
+
+  .level {
+    position: "absolute";
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+
+    &:first-child {
+      padding-top: 0;
+    }
+
+    &:last-child {
+      padding-bottom: 0;
+    }
+  }
+}
+
+.footer {
+  justify-content: space-between;
+}
