+++ SharePointFramework/PortfolioWebParts/src/components/ProjectProvision/ProvisionDrawer/index.ts
@@ -1,0 +1,1 @@
+export * from './ProvisionDrawer'
