+++ SharePointFramework/PortfolioWebParts/src/components/ProjectProvision/ProvisionDrawer/styles.ts
@@ -1,23 +1,5 @@
 import { makeStyles, tokens } from '@fluentui/react-components'
 
-export const useStyles = makeStyles({
-  level: {
-    position: 'absolute',
-    top: 0,
-    left: 0,
-    right: 0,
-    bottom: 0,
-
-    ':first-child': {
-      paddingTop: 0
-    },
-
-    ':last-child': {
-      paddingBottom: 0
-    }
-  }
-})
-
 export const useMotionStyles = makeStyles({
   toolbarButton: {
     opacity: 0,
