+++ SharePointFramework/PortfolioWebParts/src/components/ProjectProvision/ProvisionStatus/ProvisionStatus.module.scss
@@ -2,3 +2,7 @@
   display: flex;
   gap: 8px;
 }
+
+div[role="dialog"] {
+  min-width: 1120px;
+}
